// src/app/models/vehicle.model.ts
export class vehicleModel  {

    _id: string;
    brand: string;
    model: string;
    class: string;
    kind: string;
    service: string;
    licensePlate: string;
    color: string;
    year: number;
    chassisNumber: string;
    engineNumber: string;
    ramv: string;
    country: string;
    tonnage: number;
    cylindrical: number;
    fuel: string;
    provinceLocation: string;
    cantonLocation: string;
    transportCooperative: string;
    route: string;
    yearsServiceSecurity: number;

    constructor() {}
}