// src/app/models/financialData.model.ts
export class financialDataModel {

    _id: string;
    subtotal: number;
    iva: number;
    total: number;
    entryAmount: number;
    amountFinanced: number;
    legalCosts: number;
    deviceCosts: number;
    totalCredit: number;
    totalWithCosts: number;
    term: number;
    graceTime: number;
    accessoryCost: number;
    firstDueDate: Date;
    lastDueDate: Date;
    equityPercentage: number;
    disposableIncome: number;
    referencePercentage: number;
    endorsementDate: Date;
    interes: number;


    constructor() {}
}